UPDATED FUNCTIONALITY
This is the first iteration of the restaurant reviews app.

Bundled within are a few tweaks to the original starter code
to make the site responsive.

The first changes to the starter code includes a modification of
the main.js and restaurant_info.js files.

I have included image handlers in both files
 to render the restaurant images in the restaurant cards.

The CSS file had many elements that were poorly positioned.

I incorporated media queries to make the static page
respond to the viewport.

Finally I have updated the README file to make it easier for any party
iterested with the project to get up and running with it.