UPDATES
This commit includes a responsive interface and an active service worker.

I have added a missing ARIA label to the restaurant.html file.

The previous commit had some white space around the app.
This commit has been optimized to get rid of the whitespace.
The minimum width for proper scaling for this web app is 295px.